requirements_scraper.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
schedule==1.2.0
neo4j==5.14.1
python-dotenv==1.0.0
openai==1.3.0
langchain==0.1.0